♻️  Refact: 클래스와 뷰의 계층 구조에 따른 파일 분리
